+++ flowman-spec/src/test/scala/com/dimajix/flowman/spec/target/RelationTargetTest.scala
@@ -30,7 +30,9 @@ import org.scalatest.matchers.should.Matchers
 import com.dimajix.common.No
 import com.dimajix.common.Unknown
 import com.dimajix.common.Yes
+import com.dimajix.flowman.execution.BuildPolicy
 import com.dimajix.flowman.execution.Context
+import com.dimajix.flowman.execution.OutputMode
 import com.dimajix.flowman.execution.Phase
 import com.dimajix.flowman.execution.Session
 import com.dimajix.flowman.execution.Status
@@ -512,5 +514,202 @@ class RelationTargetTest extends AnyFlatSpec with Matchers with MockFactory with
         (relation.loaded _).expects(*,*).returns(No)
         (relation.exists _).expects(*).returns(No)
         target.execute(executor, Phase.VERIFY).status should be(Status.FAILED)
+    }
+
+    it should "support buildPolicy ALWAYS" in {
+        val relationGen = mock[Prototype[Relation]]
+        val relation = mock[Relation]
+        val project = Project(
+            name = "test",
+            relations = Map("relation" -> relationGen)
+        )
+
+        val session = Session.builder()
+            .withSparkSession(spark)
+            .withProject(project)
+            .withConfig("flowman.default.target.verifyPolicy", "EMPTY_AS_SUCCESS_WITH_ERRORS")
+            .build()
+        val executor = session.execution
+        val context = session.getContext(project)
+        (relationGen.instantiate _).expects(context, None).returns(relation)
+
+        val target1 = RelationTarget(
+            context,
+            RelationIdentifier("relation"),
+            MappingOutputIdentifier("mapping")
+        ).copy(buildPolicy = BuildPolicy.ALWAYS)
+        target1.dirty(executor, Phase.BUILD) should be (Yes)
+        target1.dirty(executor, Phase.BUILD) should be(Yes)
+        target1.dirty(executor, Phase.BUILD) should be(Yes)
+
+        val target2 = RelationTarget(
+            context,
+            RelationIdentifier("relation"),
+            MappingOutputIdentifier("mapping")
+        ).copy(
+            buildPolicy = BuildPolicy.ALWAYS,
+            partition = Map("part" -> "0")
+        )
+        target2.dirty(executor, Phase.BUILD) should be(Yes)
+        target2.dirty(executor, Phase.BUILD) should be(Yes)
+        target2.dirty(executor, Phase.BUILD) should be(Yes)
+    }
+
+    it should "support buildPolicy COMPAT" in {
+        val relationGen = mock[Prototype[Relation]]
+        val relation = mock[Relation]
+        val project = Project(
+            name = "test",
+            relations = Map("relation" -> relationGen)
+        )
+
+        val session = Session.builder()
+            .withSparkSession(spark)
+            .withProject(project)
+            .withConfig("flowman.default.target.verifyPolicy", "EMPTY_AS_SUCCESS_WITH_ERRORS")
+            .build()
+        val executor = session.execution
+        val context = session.getContext(project)
+        (relationGen.instantiate _).expects(context, None).returns(relation)
+
+        val target1 = RelationTarget(
+            context,
+            RelationIdentifier("relation"),
+            MappingOutputIdentifier("mapping")
+        ).copy(buildPolicy = BuildPolicy.COMPAT)
+
+        (relation.loaded _).expects(*, *).returns(Yes)
+        target1.dirty(executor, Phase.BUILD) should be(No)
+        (relation.loaded _).expects(*, *).returns(No)
+        target1.dirty(executor, Phase.BUILD) should be(Yes)
+        (relation.loaded _).expects(*, *).returns(Unknown)
+        target1.dirty(executor, Phase.BUILD) should be(Unknown)
+
+        val target2 = RelationTarget(
+            context,
+            RelationIdentifier("relation"),
+            MappingOutputIdentifier("mapping")
+        ).copy(
+            buildPolicy = BuildPolicy.COMPAT,
+            partition = Map("part" -> "0")
+        )
+
+        (relation.loaded _).expects(*, *).returns(Yes)
+        target2.dirty(executor, Phase.BUILD) should be(No)
+        (relation.loaded _).expects(*, *).returns(No)
+        target2.dirty(executor, Phase.BUILD) should be(Yes)
+        (relation.loaded _).expects(*, *).returns(Unknown)
+        target2.dirty(executor, Phase.BUILD) should be(Unknown)
+
+        val target3 = RelationTarget(
+            context,
+            RelationIdentifier("relation"),
+            MappingOutputIdentifier("mapping")
+        ).copy(mode = OutputMode.APPEND)
+
+        target3.dirty(executor, Phase.BUILD) should be(Yes)
+        target3.dirty(executor, Phase.BUILD) should be(Yes)
+        target3.dirty(executor, Phase.BUILD) should be(Yes)
+    }
+
+    it should "support buildPolicy IF_EMPTY" in {
+        val relationGen = mock[Prototype[Relation]]
+        val relation = mock[Relation]
+        val project = Project(
+            name = "test",
+            relations = Map("relation" -> relationGen)
+        )
+
+        val session = Session.builder()
+            .withSparkSession(spark)
+            .withProject(project)
+            .withConfig("flowman.default.target.verifyPolicy", "EMPTY_AS_SUCCESS_WITH_ERRORS")
+            .build()
+        val executor = session.execution
+        val context = session.getContext(project)
+        (relationGen.instantiate _).expects(context, None).returns(relation)
+
+        val target1 = RelationTarget(
+            context,
+            RelationIdentifier("relation"),
+            MappingOutputIdentifier("mapping")
+        ).copy(buildPolicy = BuildPolicy.IF_EMPTY)
+
+        (relation.loaded _).expects(*, *).returns(Yes)
+        target1.dirty(executor, Phase.BUILD) should be(No)
+        (relation.loaded _).expects(*, *).returns(No)
+        target1.dirty(executor, Phase.BUILD) should be(Yes)
+        (relation.loaded _).expects(*, *).returns(Unknown)
+        target1.dirty(executor, Phase.BUILD) should be(Unknown)
+
+        val target2 = RelationTarget(
+            context,
+            RelationIdentifier("relation"),
+            MappingOutputIdentifier("mapping")
+        ).copy(
+            buildPolicy = BuildPolicy.IF_EMPTY,
+            partition = Map("part" -> "0")
+        )
+
+        (relation.loaded _).expects(*, *).returns(Yes)
+        target2.dirty(executor, Phase.BUILD) should be(No)
+        (relation.loaded _).expects(*, *).returns(No)
+        target2.dirty(executor, Phase.BUILD) should be(Yes)
+        (relation.loaded _).expects(*, *).returns(Unknown)
+        target2.dirty(executor, Phase.BUILD) should be(Unknown)
+    }
+
+    it should "support buildPolicy SMART" in {
+        val relationGen = mock[Prototype[Relation]]
+        val relation = mock[Relation]
+        val project = Project(
+            name = "test",
+            relations = Map("relation" -> relationGen)
+        )
+
+        val session = Session.builder()
+            .withSparkSession(spark)
+            .withProject(project)
+            .withConfig("flowman.default.target.verifyPolicy", "EMPTY_AS_SUCCESS_WITH_ERRORS")
+            .build()
+        val executor = session.execution
+        val context = session.getContext(project)
+        (relationGen.instantiate _).expects(context, None).returns(relation)
+
+        val target1 = RelationTarget(
+            context,
+            RelationIdentifier("relation"),
+            MappingOutputIdentifier("mapping")
+        ).copy(buildPolicy = BuildPolicy.SMART)
+
+        target1.dirty(executor, Phase.BUILD) should be(Yes)
+        target1.dirty(executor, Phase.BUILD) should be(Yes)
+        target1.dirty(executor, Phase.BUILD) should be(Yes)
+
+        val target2 = RelationTarget(
+            context,
+            RelationIdentifier("relation"),
+            MappingOutputIdentifier("mapping")
+        ).copy(
+            buildPolicy = BuildPolicy.SMART,
+            partition = Map("part" -> "0")
+        )
+
+        (relation.loaded _).expects(*, *).returns(Yes)
+        target2.dirty(executor, Phase.BUILD) should be(No)
+        (relation.loaded _).expects(*, *).returns(No)
+        target2.dirty(executor, Phase.BUILD) should be(Yes)
+        (relation.loaded _).expects(*, *).returns(Unknown)
+        target2.dirty(executor, Phase.BUILD) should be(Unknown)
+
+        val target3 = RelationTarget(
+            context,
+            RelationIdentifier("relation"),
+            MappingOutputIdentifier("mapping")
+        ).copy(mode = OutputMode.APPEND)
+
+        target3.dirty(executor, Phase.BUILD) should be(Yes)
+        target3.dirty(executor, Phase.BUILD) should be(Yes)
+        target3.dirty(executor, Phase.BUILD) should be(Yes)
     }
 }