+++ docs/releases.md
@@ -31,6 +31,8 @@ changes over time.
 * github-295: Create build profile for CDP 7.1 with Spark 3.2
 * github-296: Update npm dependencies (vuetify & co)
 * github-297: Parametrize when to execute a specific phase
+* github-299: Move migrationPolicy and migrationStrategy from target into relation
+* github-115: Implement additional build policy in relation target for forcing dirty
 
 
 ### Version 0.29.0 - 2022-11-08